style: 변수명, 훅스명 명시적으로 변경
